after updraft

+++ web/wp-content/plugins/wp-optimize/minify/class-wp-optimize-minify-commands.php
@@ -32,7 +32,9 @@ class WP_Optimize_Minify_Commands {
 	public function purge_all_minify_cache() {
 		if (!WPO_MINIFY_PHP_VERSION_MET) return array('error' => __('WP-Optimize Minify requires a higher PHP version', 'wp-optimize'));
 		WP_Optimize_Minify_Cache_Functions::purge();
+		WP_Optimize_Minify_Cache_Functions::cache_increment();
 		$others = WP_Optimize_Minify_Cache_Functions::purge_others();
+		$files = $this->get_minify_cached_files();
 		$message = array(
 			__('The minification cache was deleted.', 'wp-optimize'),
 			strip_tags($others, '<strong>'),
@@ -40,7 +42,8 @@ class WP_Optimize_Minify_Commands {
 		$message = array_filter($message);
 		return array(
 			'success' => true,
-			'message' => implode("\n", $message)
+			'message' => implode("\n", $message),
+			'files' => $files
 		);
 	}
 
@@ -52,8 +55,10 @@ class WP_Optimize_Minify_Commands {
 	public function minify_increment_cache() {
 		if (!WPO_MINIFY_PHP_VERSION_MET) return array('error' => __('WP-Optimize Minify requires a higher PHP version', 'wp-optimize'));
 		WP_Optimize_Minify_Cache_Functions::cache_increment();
+		$files = $this->get_minify_cached_files();
 		return array(
-			'success' => true
+			'success' => true,
+			'files' => $files
 		);
 	}
 
@@ -64,11 +69,20 @@ class WP_Optimize_Minify_Commands {
 	 */
 	public function purge_minify_cache() {
 		if (!WPO_MINIFY_PHP_VERSION_MET) return array('error' => __('WP-Optimize Minify requires a higher PHP version', 'wp-optimize'));
+		if (!WP_Optimize()->can_purge_the_cache()) return array('error' => __('You do not have permission to purge the cache', 'wp-optimize'));
+
 		// deletes temp files and old caches incase CRON isn't working
 		WP_Optimize_Minify_Cache_Functions::cache_increment();
-		$state = WP_Optimize_Minify_Cache_Functions::purge_temp_files();
-		$old = WP_Optimize_Minify_Cache_Functions::purge_old();
+		if (wp_optimize_minify_config()->always_purge_everything()) {
+			WP_Optimize_Minify_Cache_Functions::purge();
+			$state = array();
+			$old = array();
+		} else {
+			$state = WP_Optimize_Minify_Cache_Functions::purge_temp_files();
+			$old = WP_Optimize_Minify_Cache_Functions::purge_old();
+		}
 		$others = WP_Optimize_Minify_Cache_Functions::purge_others();
+		$files = $this->get_minify_cached_files();
 
 		$notice = array(
 			__('All caches from WP-Optimize Minify have been purged.', 'wp-optimize'),
@@ -82,7 +96,8 @@ class WP_Optimize_Minify_Commands {
 			'others' => $others,
 			'state' => $state,
 			'message' => $notice,
-			'old' => $old
+			'old' => $old,
+			'files' => $files
 		);
 	}
 
@@ -105,6 +120,20 @@ class WP_Optimize_Minify_Commands {
 			}
 		}
 
+		if (isset($data['minify_advanced_tab'])) {
+			// Make sure that empty settings are still saved
+			if (!isset($new_data['ignore_list'])) $new_data['ignore_list'] = array();
+			if (!isset($new_data['blacklist'])) $new_data['blacklist'] = array();
+		}
+
+		/**
+		 * Filters the data before saving it
+		 *
+		 * @param array $new_data - The original data
+		 * @return array The data, altered or not
+		 */
+		$new_data = apply_filters('wpo_save_minify_settings', $new_data);
+
 		if (!class_exists('WP_Optimize_Minify_Config')) return array(
 			'success' => false,
 			'message' => "WP_Optimize_Minify_Config class doesn't exist",
@@ -116,10 +145,10 @@ class WP_Optimize_Minify_Commands {
 				'error' => 'failed to save'
 			);
 		}
-		WP_Optimize_Minify_Cache_Functions::cache_increment();
-		WP_Optimize_Minify_Cache_Functions::purge_others();
+		$purged = $this->purge_minify_cache();
 		return array(
-			'success' => true
+			'success' => true,
+			'files' => $purged['files']
 		);
 	}
